Benchmark cleanup
