+++ benches/global_criterion_settings.rs
@@ -1,0 +1,5 @@
+use criterion::{Criterion};
+
+pub fn get_criterion() -> Criterion {
+    Criterion::default().noise_threshold(0.10)
+}
